feat : add input wrapping class and move some logic to it
